Add a babel build
